+++ NEWS.md
@@ -1,3 +1,9 @@
+# deepdep 0.4.1
+* fixed bug where `plot_dependencies()` wouldn't remove all invalid dependencies when using `declutter = TRUE` (issue #35)
+* fixed bug where `plot_dependencies()` would sometimes raise an error when using `show_version = TRUE` and `depth >= 2` (issue #36)
+* fixed bug where `plot_dependencies()` would plot one non-central label even when using `label_percentage = 0` (issue #38)
+* tests that use CRAN download counts now skip on CRAN only
+
 # deepdep 0.4.0
 * removed hard dependencies on plotting-related packages; `ggplot2`, `ggraph`, `graphlayouts`, `igraph`, and `scales` are now Suggests instead of Imports (issue #32)
 * removed unused dependency on `ggforce`